Convert to TypeScript
This is a first pass/spike at converting the template to TypeScript. I did this
since I wanted to see how it would work compared to my recent experience with
Flow.

I'm planning on also putting together a Flow one for comparison.

So far, my experience setting this up was smoother than with Flow. Some of that
could be getting more used to how defining types in JS works, but the tooling
definitely felt nicer with TypeScript, even without using autocomplete. The
errors seemed more reliable and it didn't freak out as much as Flow has on me.

It's very possible this is due to this being a much simpler implementation than
the project I'm on, so I can't make much of a definitive opinion yet on which
would be preferred.